requirements.txt
streamlit==1.32.0
tensorflow-cpu==2.15.0
numpy==1.24.3
Pillow==10.2.0 